requirements.txt
Flask==2.3.3
requests==2.31.0
Pillow==10.1.0
python-dotenv==1.0.1
gunicorn==21.2.0
